DbAccess v2

+++ Clickfly/ViewModels/FlightSubtotalRequest.cs
@@ -4,7 +4,7 @@ namespace clickfly.ViewModels
 {
     public class FlightSubtotalRequest
     {
-        public double selected_seats { get; set; }
+        public int selected_seats { get; set; }
         public string flight_segment_id { get; set; }
     }
 }
